The troubles with veil and train drift by And i to my heart, obnoxious comet cry The dresses swap to dust of mortality Ready to be judge under that stern law A loud noise as if in the thunders of the dark Thrills in ears as a sign of commencement.
The pains burns the minds in memory of you
whiles it burns the soul alone in hell
since the first death, can not out-freeze our night
where is that fire of love that will warm our hands
when our sins are bundle on our head.

oh! pillars of flame that drift across my heart,
The eyes are now seas of fire
The tears are rolling falling walls of true
All things underneath darkness are faced with light.
Even the feet that used to elude like rabbit
And the mouth that used to vehemence in lie
the mouth vehemence in lie yesterday
Are in sonorous propriety in hell.

Hell is not a lodging place for the righteous to sleep.
Instead it is a final stage for all wickedness to end.